client/requirements.txt
websockets==15.0.1
anthropic==0.70.0
python-dotenv==1.1.1
pydantic==2.12.2